Rewrite Part 1 of Day 02 to use `pest_consume`
This changes from the fairly blunt `for` loop approach in the first version to a solution using the `pest_consume` package. This meant we created domain-level structures, which is good.

I find the `impl GameParser` section a wee bit tedious, although that may be mostly a matter of not being used to handling Pest grammars.

That said, It's definitely an improvement on the `for` loop solution.