Add a Day 03 Pest grammar
Thanks to @Justus_Fluegel for suggesting the nice idea of using `"."` as `WHITESPACE`.

We flailed a little on the `symbol` until I understood that negation was a "check" and not a parsing element, and that we needed the `~ ANY` to actually match a character.

Luckily, the Pest VSCode extension gave us an error/warning when we didn't have the `~ ANY` bit, which forced us to look some things up and figure out how negation really works in Pest.