Implement Day 04 parsing
This gets the parsing working for Day 04, and stubs out the `ScratchCard` type and the necessary methods.

Based on some conversation at the end, I think we want to change `winning_numbers` and `our_numbers` to be `HashSet`s so we can use `.intersection()` to compute the (number of) winning cards.