Move the parsing into `FromStr`
Having the `parse_scratchcards` in `ScratchCard` didn't really make a ton of sense, and I realized that what I really wanted was an implementation of `FromStr` like we did in most of the 2022 AoC solutions. (Not sure why I didn't realize that this year, but whatever.)

To make this work I needed to create a `ScratchCards` type  wrapping a `Vec<ScratchCard>` and implement `IntoIterator` for that type so we could iterator over the cards.