Write a Pest grammar for Day 05
I debated whether to even use a parsing tool for Day 5 since you could split into lines and then split lines into words. In the end, though, I went ahead and used Pest to parse it. @JustusFluegel really wants me to use Chumsky to parse something in Advent of Code, but I really should do a little homework on Chumsky before we try that. @MizardX suggested that Day 06 might be good for that, so maybe we can try that then (assuming I do at least a little homework).

The grammar here was pretty straightforward, although I completely forgot the `range_mapping` entry the first time. We also had some issues with whitespace that I don't fully understand. I initially had a space in the `"map: "` terminal that broke things, but it was an easy fix.

We tried using tags again, but then never used them in the parsing code. I still don't honestly understand what the role of tags is in Pest. Maybe it would make more sense if we weren't using `pest-consume`?