Implement the parser for Day 05
This gets the parsing done for Day 05. I don't know if our data structures are quite right, but until we start trying to solve the problem we won't really know. Hopefully they're reasonable, though.

Now we "just" need to solve the problem. Part 1 looks like it could just be a pretty straightforward grind through the data, but apparently Part 2 is a bear so we might need to make some changes there.