Solve Part 2 of Day 05
We basically went the brute force route, with the only concession to "optimization" being using Rayon's `par_iter()` to parallelize the processing of the bazillion seeds defined by the seed ranges. This got the (correct) answer about 10 seconds on my laptop (with 12 cores).

We might come back to this next week, and process ranges as the "primitives" instead of individual seeds.