Convert `Almanac` hold an `Option<Mapping>`
We're composing all the individual `Mapping`s into a single `Mapping`, so we don't want/need `Vec<Mapping> any more.

We wrapped the composed `Mapping` in `Option` to handle the case that there are not actually any source mappings.  This was suggested by the fact that `reduce(Mapping::compose)` returns an `Option` in case the collection `maps()` is empty, and we chose to just pass that `Option` along. (Thanks to @JustusFluegel for helping provide some clarity on this during the stream.)