Replace `convert()` with `compose()` and `lookup()`
The previous version of this used `convert()` to map source values to target values. We replaced that with `compose()`, which composed various sequential mappings into a single mapping (much like function composition). `compose()` ends up being implemented for both `Mapping` and `RangeMapping`.

We also needed a `Mapping::lookup()` method that finds which `RangeMapping` in this `Mapping` contains a given `source_index`.