Unignore Day 05, Part 2 test
This removes an `#[ignore]` on the Day 05, Part 2 full input test. That took forever when we had the slow version of the program, but now that we have the sub-millisecond implementation we should include this test.

I also had the wrong module names on the Day 05 tests due to copy-paste problems, and this fixes that.