Replace `expect`s with "proper" errors in Day 03, Part 1
This replaces a couple of "expect" calls with actual Pest errors for Part 1 of Day 03. I started doing this because I was hoping it would get rid of a Clippy warning about unnecessary wraps in the Pest parser, but it doesn't affect that at all. It's sitll probably a good idea, though, so I'm glad I did it.

To get rid of the Clippy warning I had to add an `#[allow(clippy::unnecessary_wraps)]` on the whole `impl SchaticParser` block. That seems a little annoying, but there we are. That Clippy warning is new, so I suspect the Pest folks may update things to handle this more gracefully in the not-too-distant future.