Add brute-force solution to Part 2 of Day 08
This adds an unacceptably slow solution to Part 2 of Day 08. @MizardX suggests that this might take something on the order of 2^43 steps, which is clearly never finishing, so we'll need to look for an algorithmically smarter solution.

The main change was moving from having the current "state" be a single node to being a `Vec` of nodes, all of which get updated.

We also had to make a change to our parsing since their new test input had digits in the node names, which we hadn't supported initially.