Complete Part 2 of Day 08
We got it to work! But it took an entire two-hour streaming session. Lots of good suggestions from @MizardX and Vincent_SIFTD@Twitch.

In the end, after a lot of poking about we determined that there were essentially 6 independent graphs, each of which took a different starting node to a different ending node. Each path was a (different) multiple of 271 (the original path length), which is prime.