Factor out duplicate code in Day 11, Part 2
There was a nasty bit of code duplication in Part 2 of Day 11, that I actually got burned on at least once and Vincent_SIFTD@Twitch also wanted to see refactored.

The two code blocks differed solely in whether they returned a row or a column, so it seemed that this would be easy, but it turned out to be a bit of a mess, with some *very* ugly types.

We did it, using an implementation of `Fn` and `FnMut` to provide the an immu