Simple (parallel) solution to Day 12, Part 2
This (I think) solves Part 2 of Day 12, but it's super slow. It does pass the test case, so I'm actually reasonably confident that the logic is right, but the run time is too slow.

We added parallelism with (`par_iter()`), and that probably sped things up considerably, but I'm guessing that we're still in hours to get this done.

We really need to use memoization or dynamic programming. We tried slapping the `memoize` crate on the problem, but that didn't work because it can't be applied to methods. Next week we'll try to restructure things so we can use that crate, and hopefully that will speed things up enough to get us through the problem.