Set up Day 13
We added the `ndarray` crate to handle the management of the 2D array of locations in each pattern.

We then implemented all the parsing into an `ndarray` `array2`, and that worked out very nicely.

This also adds the test files and sets up the test code using the new types.