Implement Part 1, Day 13
This was both "easy" and not. I really liked using `ndarray`; it did a lot of nice things for us, but we also spent a lot of time figuring out what it could/would do for us.

The big (eventual) win was using the idea of lanes and axes correctly. Once we had that sorted out, then everything was actually pretty easy, and we managed to avoid pretty much any interesting code duplication. There was a lot of looking things up and refactoring to get there, though.

Huge thanks to @JustusFluegel and @MizardX for a ton of important suggestions in getting here.

There were a lot of nice uses of iterator tools like `filter_map`, `find`, `find_map`, and `zip` here.