Implemented Day 13, Part 13
This added the ability to "smudge in place" to flip locations in a side-effecting way.

We added a `Reflection` type to keep track of the reflections that we had discovered. We also added a `Reflection::value` method that captured the multiplier logic.

The key change to the logic was to add `reflection_value_with_smudges` that loops over all the locations, smudges them, and checks for differences.

We also replaced all the `find` logic with things that returned collections of values so that we could use set difference to find the new reflections.