Add hash code tracking for Day 14 Part 2
@MizardX suggested tracking hash codes of `Platform` arrays as a way of efficiently figuring out when we'd found a cycle in our sequence of `Platform`s as we tilted them.

This is a quick hack just to see if that gets us timely termination, and the answer is a resounding "Yes!" as we can find the cycle in under a second. Then all that's left is to figure out where in the cycle the 1 billionth value will land, and return that.