Set up types and parsing for Day 15, Part 2
This adds a bunch of new types for Day 15, Part 2 (with some nice use of Rust's type system), and updates the parsing to build these types as appropriate.

In parsing into a `Step`, we used a really cool *spread* trick to match against a slice pattern. I don't think I'd ever used that, and thanks to @MizardX for the suggestion.

We also had to introduce an actual error type (`ParseStepError`) for that parsing since there were now ways that parsing could fail.

We're going to be using a `HashMap<Label, Vec<Lens>>`, so we needed to move the `impl Hash` to `Lens`, and add a custom `impl PartialEq` for `Lens`. We "cheated" and just had `eq()` return `true` all the time, counting on `HashMap` to discriminate different `Label`s based on their hash values.