Insert lenses into boxes
This builds the list of boxes and lenses from the list of instructions.

We had to derive several new traits.

We had to replace our hack for `PartialEq for Label` because that didn't work. It turns out that we needed to actually compare the hash values or everything ended up in the same bucket in the hash table.

We changed several things to use references to labels to avoid a bunch of label cloning.