Implement Day 16, Part 2
We just used brute force, and it turned to work quite reasonably, completing in 15 ms and not using a huge amount of heap (<2MB).

There are numerous ways we could probably speed this up, e.g., by moving the mutable state out of the grid (so we don't have to clone the grid over and over) and/or parallelizing the `max()` computation with rayon. But it's not clear that we need to do any of this.